remove crypto.js and use services.js instead
